Create linera-wallet-json crate
Extract wallet persistence, keystore, path resolution, and display
code from linera-service into a dedicated crate.
