+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "linera-summary",
     "linera-views",
     "linera-views-derive",
+    "linera-wallet-json",
     "linera-witty",
     "linera-witty-macros",
     "linera-witty/test-modules",
@@ -60,6 +61,7 @@ default-members = [
     "linera-summary",
     "linera-views",
     "linera-views-derive",
+    "linera-wallet-json",
     "linera-witty",
     "linera-witty-macros",
     "linera-witty/test-modules",
@@ -335,6 +337,7 @@ linera-storage-service = { version = "0.16.0", path = "./linera-storage-service"
 linera-version = { version = "0.16.0", path = "./linera-version" }
 linera-views = { version = "0.16.0", path = "./linera-views", default-features = false }
 linera-views-derive = { version = "0.16.0", path = "./linera-views-derive" }
+linera-wallet-json = { version = "0.16.0", path = "./linera-wallet-json" }
 linera-witty = { version = "0.16.0", path = "./linera-witty" }
 linera-witty-macros = { version = "0.16.0", path = "./linera-witty-macros" }
