+++ linera-service/src/cli/main.rs
@@ -63,7 +63,7 @@ use linera_execution::committee::Committee;
 use linera_faucet_server::{FaucetConfig, FaucetService};
 #[cfg(with_metrics)]
 use linera_metrics::monitoring_server;
-use linera_persistent::{self as persistent, Persist, PersistExt as _};
+use linera_persistent::{self as persistent, Persist as _};
 use linera_service::{
     cli::{
         command::{
@@ -162,7 +162,7 @@ impl Runnable for Job {
     {
         let Job(options) = self;
         let mut wallet = options.wallet()?;
-        let mut signer = options.signer()?;
+        let mut keystore = options.keystore()?;
 
         let command = options.command.clone();
 
@@ -174,7 +174,7 @@ impl Runnable for Job {
                 amount,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_client = context.make_chain_client(sender.chain_id).await?;
                 info!(
@@ -204,10 +204,12 @@ impl Runnable for Job {
                 balance,
                 super_owner,
             } => {
-                let new_owner = owner.unwrap_or_else(|| signer.generate_new().into());
-                signer.persist().await?;
+                let new_owner = match owner {
+                    Some(owner) => owner,
+                    None => keystore.generate_key().await?.into(),
+                };
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -254,7 +256,7 @@ impl Runnable for Job {
                 application_permissions_config,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -299,7 +301,7 @@ impl Runnable for Job {
 
             ShowOwnership { chain_id } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let ownership = context.ownership(chain_id).await?;
                 let json = serde_json::to_string_pretty(&ownership)?;
@@ -311,14 +313,14 @@ impl Runnable for Job {
                 ownership_config,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 context.change_ownership(chain_id, ownership_config).await?
             }
 
             SetPreferredOwner { chain_id, owner } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 context.set_preferred_owner(chain_id, owner).await?
             }
@@ -328,7 +330,7 @@ impl Runnable for Job {
                 application_permissions_config,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -359,7 +361,7 @@ impl Runnable for Job {
 
             CloseChain { chain_id } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_client = context.make_chain_client(chain_id).await?;
                 info!("Closing chain {}", chain_id);
@@ -394,7 +396,7 @@ impl Runnable for Job {
 
             LocalBalance { account } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let account = account.unwrap_or_else(|| context.default_account());
                 let chain_client = context.make_chain_client(account.chain_id).await?;
@@ -408,7 +410,7 @@ impl Runnable for Job {
 
             QueryBalance { account } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let account = account.unwrap_or_else(|| context.default_account());
                 let chain_client = context.make_chain_client(account.chain_id).await?;
@@ -425,7 +427,7 @@ impl Runnable for Job {
 
             SyncBalance { account } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let account = account.unwrap_or_else(|| context.default_account());
                 let chain_client = context.make_chain_client(account.chain_id).await?;
@@ -450,7 +452,7 @@ impl Runnable for Job {
                 until_block_time,
             } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -473,7 +475,7 @@ impl Runnable for Job {
 
             ProcessInbox { chain_id } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let follow_only = context
@@ -500,7 +502,7 @@ impl Runnable for Job {
 
             QueryShardInfo { chain_id } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 println!("Querying validators for shard information about chain {chain_id}.\n");
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -539,7 +541,7 @@ impl Runnable for Job {
 
                 let time_start = Instant::now();
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 // ResourceControlPolicy doesn't need version checks
@@ -705,7 +707,7 @@ impl Runnable for Job {
                 info!("Starting operations to remove old committees");
                 let time_start = Instant::now();
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let chain_client = context
@@ -784,13 +786,10 @@ impl Runnable for Job {
                             ..Default::default()
                         };
 
-                        let pub_keys: Vec<_> = std::iter::repeat_with(|| signer.generate_new())
-                            .take(num_chains)
-                            .collect();
-                        signer.persist().await?;
+                        let pub_keys = keystore.generate_keys(num_chains).await?;
 
                         let mut context = options
-                            .create_client_context(storage.clone(), wallet, signer.into_value())
+                            .create_client_context(storage.clone(), wallet, keystore)
                             .await?;
                         let chain_clients = context
                             .prepare_for_benchmark(
@@ -1202,7 +1201,7 @@ impl Runnable for Job {
 
             Watch { chain_id, raw } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let mut join_set = JoinSet::new();
@@ -1228,7 +1227,7 @@ impl Runnable for Job {
                 query,
             } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id
                     .or_else(|| context.wallet().default_chain())
@@ -1270,7 +1269,7 @@ impl Runnable for Job {
                 pause,
             } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let default_chain = context.wallet().default_chain();
@@ -1386,7 +1385,7 @@ impl Runnable for Job {
                 let genesis_config = wallet.genesis_config().clone();
 
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let chain_id = if let Some(chain_id) = chain_id {
@@ -1429,7 +1428,7 @@ impl Runnable for Job {
                 publisher,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let start_time = Instant::now();
@@ -1452,7 +1451,7 @@ impl Runnable for Job {
                 start_index,
             } => {
                 let context = options
-                    .create_client_context(storage.clone(), wallet, signer.into_value())
+                    .create_client_context(storage.clone(), wallet, keystore)
                     .await?;
                 let start_time = Instant::now();
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
@@ -1468,7 +1467,7 @@ impl Runnable for Job {
                 publisher,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let start_time = Instant::now();
@@ -1486,7 +1485,7 @@ impl Runnable for Job {
             // TODO(#2490): Consider removing or renaming this.
             ReadDataBlob { hash, reader } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let start_time = Instant::now();
@@ -1507,7 +1506,7 @@ impl Runnable for Job {
                 required_application_ids,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let start_time = Instant::now();
@@ -1559,7 +1558,7 @@ impl Runnable for Job {
                 required_application_ids,
             } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
 
                 let start_time = Instant::now();
@@ -1601,7 +1600,7 @@ impl Runnable for Job {
 
             Assign { owner, chain_id } => {
                 let mut context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let start_time = Instant::now();
                 info!(
@@ -1628,7 +1627,7 @@ impl Runnable for Job {
                     required_application_ids,
                 } => {
                     let mut context = options
-                        .create_client_context(storage, wallet, signer.into_value())
+                        .create_client_context(storage, wallet, keystore)
                         .await?;
                     let start_time = Instant::now();
                     let publisher = publisher.unwrap_or_else(|| context.default_chain());
@@ -1677,7 +1676,7 @@ impl Runnable for Job {
 
             RetryPendingBlock { chain_id } => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let start_time = Instant::now();
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
@@ -1708,7 +1707,7 @@ impl Runnable for Job {
                 set_default,
             }) => {
                 let start_time = Instant::now();
-                let owner: AccountOwner = signer.mutate(|s| s.generate_new()).await?.into();
+                let owner: AccountOwner = keystore.generate_key().await?.into();
 
                 info!(
                     "Requesting a new chain for owner {owner} using the faucet at address \
@@ -1736,7 +1735,7 @@ impl Runnable for Job {
                 }
 
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_client = context.make_chain_client(description.id()).await?;
                 chain_client.synchronize_from_validators().await?;
@@ -1764,7 +1763,7 @@ impl Runnable for Job {
                     anyhow::bail!("Missing network description");
                 };
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let faucet = cli_wrappers::Faucet::new(faucet_url);
                 let committee = faucet.current_committee().await?;
@@ -1779,7 +1778,7 @@ impl Runnable for Job {
 
             Wallet(WalletCommand::FollowChain { chain_id, sync }) => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let start_time = Instant::now();
                 context
@@ -1802,7 +1801,7 @@ impl Runnable for Job {
 
             Chain(ChainCommand::ShowBlock { chain_id, height }) => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_state_view = context
@@ -1824,7 +1823,7 @@ impl Runnable for Job {
 
             Chain(ChainCommand::ShowChainDescription { chain_id }) => {
                 let context = options
-                    .create_client_context(storage, wallet, signer.into_value())
+                    .create_client_context(storage, wallet, keystore)
                     .await?;
                 let chain_id = chain_id.unwrap_or_else(|| context.default_chain());
                 let chain_client = context.make_chain_client(chain_id).await?;
@@ -1843,7 +1842,7 @@ impl Runnable for Job {
                 validator_command
                     .run(
                         &mut options
-                            .create_client_context(storage, wallet, signer.into_value())
+                            .create_client_context(storage, wallet, keystore)
                             .await?,
                     )
                     .await?;
@@ -2230,8 +2229,8 @@ async fn run(options: &Options) -> Result<i32, Error> {
                 Timestamp::from(micros)
             });
 
-            let mut signer = options.create_keystore(*testing_prng_seed)?;
-            let admin_public_key = signer.mutate(|s| s.generate_new()).await?;
+            let mut keystore = options.create_keystore(*testing_prng_seed)?;
+            let admin_public_key = keystore.generate_key().await?;
 
             let network_name = network_name.clone().unwrap_or_else(|| {
                 // Default: e.g. "linera-2023-11-14T23:13:20"
@@ -2260,7 +2259,7 @@ async fn run(options: &Options) -> Result<i32, Error> {
             )];
             for _ in 0..*num_other_initial_chains {
                 // Create keys.
-                let public_key = signer.mutate(|s| s.generate_new()).await?;
+                let public_key = keystore.generate_key().await?;
                 let description = genesis_config.add_root_chain(public_key, *initial_funding);
                 let chain = wallet::Chain {
                     owner: Some(public_key.into()),
@@ -2316,8 +2315,8 @@ async fn run(options: &Options) -> Result<i32, Error> {
 
         ClientCommand::Keygen => {
             let start_time = Instant::now();
-            let mut signer = options.signer()?;
-            let public_key = signer.mutate(|s| s.generate_new()).await?;
+            let mut keystore = options.keystore()?;
+            let public_key = keystore.generate_key().await?;
             let owner = AccountOwner::from(public_key);
             println!("{}", owner);
             info!("Key generated in {} ms", start_time.elapsed().as_millis());
@@ -2456,7 +2455,7 @@ async fn run(options: &Options) -> Result<i32, Error> {
                         println!("{chain_id}");
                     }
                 } else {
-                    wallet.pretty_print(chain_ids);
+                    linera_wallet_json::display::pretty_print(&wallet, chain_ids);
                 }
                 Ok(0)
             }
@@ -2475,7 +2474,7 @@ async fn run(options: &Options) -> Result<i32, Error> {
                 let start_time = Instant::now();
                 let owner = options.wallet()?.forget_keys(*chain_id)?;
                 if !options
-                    .signer()?
+                    .keystore()?
                     .contains_key(&owner)
                     .await
                     .expect("Signer error")
@@ -2534,7 +2533,7 @@ Make sure to use a Linera client compatible with this network.
                     }
                 };
                 let mut keystore = options.create_keystore(*testing_prng_seed)?;
-                keystore.persist().await?;
+                keystore.save().await?;
                 options.create_wallet(genesis_config)?.save()?;
                 options.initialize_storage().boxed().await?;
                 options.run_with_storage(Job(options.clone())).await??;
