+++ linera-service/src/cli/options.rs
@@ -4,10 +4,8 @@
 use std::path::PathBuf;
 
 use anyhow::Error;
-use linera_base::crypto::InMemorySigner;
 use linera_client::{client_context::ClientContext, config::GenesisConfig};
 use linera_execution::WithWasmDefault as _;
-use linera_persistent as persistent;
 use linera_service::{
     cli::{command::ClientCommand, common_options::CommonCliOptions},
     storage::{Runnable, RunnableWithStore, StorageConfig},
@@ -162,8 +160,8 @@ impl Options {
         self.common.wallet()
     }
 
-    pub fn signer(&self) -> Result<persistent::File<InMemorySigner>, Error> {
-        self.common.signer()
+    pub fn keystore(&self) -> Result<linera_wallet_json::Keystore, Error> {
+        self.common.keystore()
     }
 
     pub fn create_wallet(&self, genesis_config: GenesisConfig) -> Result<Wallet, Error> {
@@ -173,7 +171,7 @@ impl Options {
     pub fn create_keystore(
         &self,
         testing_prng_seed: Option<u64>,
-    ) -> Result<persistent::File<InMemorySigner>, Error> {
+    ) -> Result<linera_wallet_json::Keystore, Error> {
         self.common.create_keystore(testing_prng_seed)
     }
 }
