+++ linera-service/src/lib.rs
@@ -17,5 +17,4 @@ pub mod storage;
 pub mod task_processor;
 pub mod tracing;
 pub mod util;
-pub mod wallet;
-pub use wallet::Wallet;
+pub use linera_wallet_json::PersistentWallet as Wallet;
