+++ linera-wallet-json/README.md
@@ -1,0 +1,3 @@
+# linera-wallet-json
+
+Persistent wallet implementation for the Linera protocol.
