+++ linera-wallet-json/src/lib.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod display;
+pub mod keystore;
+pub mod paths;
+pub mod wallet;
+
+pub use keystore::Keystore;
+pub use wallet::PersistentWallet;
