+++ packages.txt
@@ -14,6 +14,7 @@ linera-storage-service
 linera-cache
 linera-storage
 linera-core
+linera-wallet-json
 linera-ethereum
 linera-metrics
 linera-sdk-derive
